requirements.txt
requests
pylint 
autopep8
python-dotenv
together
flask